% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_french.R
\name{fixed_french}
\alias{fixed_french}
\title{Fixed-rate french loanamortization schedule calculation function.}
\usage{
fixed_french(
  amount,
  interest_rate,
  start_date,
  end_date = NULL,
  n_payments = NULL,
  frequency = 12,
  optional_parameters = NULL,
  truncate_month_day = TRUE
)
}
\arguments{
\item{amount}{initial principal amount}

\item{interest_rate}{yearly interest rate}

\item{start_date}{date of first loan payment}

\item{end_date}{date in which the last loan payment is expected. This argument overrides n_payments.}

\item{n_payments}{number of payments expected.}

\item{frequency}{number of months in a year in which payments are expected}

\item{optional_parameters}{named vector containing constant columns to be added to the return dataframe}

\item{truncate_month_day}{if true, all the days in the return dataframe dates column will be set to zero}
}
\value{
dataframe containing the amortization schedule for the loan described in the arguments
}
\description{
Fixed-rate french loanamortization schedule calculation function.
}
\examples{
fixed_french(amount=10000, interest_rate=0.01, start_date=as.Date("2020-03-19"), end_date=as.Date("2032-03-19"), frequency=4, optional_parameters=c("loan_id"="123"), truncate_month_day=FALSE)
add(10, 1)
}
